Move Trait set objects into their own module.
